Cleanup code via IntelliJ's code inspection.
Replace String.indexOf() with String.contains().
replace StringBuffer with StringBuilder.
Remove unnecessary boxing.
Replace explicit Generics type with <>.
Collapse catch blocks.
Simplify via lambdas.
Simplify Map operations.
Use Objects.requireNonNull().

+++ mucommander-protocol-nfs/src/main/java/com/sun/gssapi/DERParser.java
@@ -113,8 +113,8 @@ class DERParser {
 	    else
 		comp = 2;	
 				
-	    v.addElement(Integer.valueOf(comp));
-	    v.addElement(Integer.valueOf(tmp - (40 * comp)));
+	    v.addElement(comp);
+	    v.addElement(tmp - (40 * comp));
 			
 	    //get the rest of the octets
 	    for (int i = 1; i < numOfOctets; i++) {
@@ -129,7 +129,7 @@ class DERParser {
 			break;
 		    i++;
 		}
-		v.addElement(Integer.valueOf(comp));
+		v.addElement(comp);
 	    }
 		
 	} catch (IOException e) {
